server/requirements.txt
Django==5.1.1
djangorestframework==3.15.1
django-cors-headers==4.4.0
